Refactor: extract macro runtime classes to shared files
