+++ macro/src/main/scala/org/mockito/WhenMacro.scala
@@ -2,13 +2,16 @@ package org.mockito
 
 import org.mockito.Utils.*
 import org.mockito.internal.MacroDebug.debugResult
-import org.mockito.internal.ValueClassWrapper
 import org.mockito.stubbing.{ ScalaFirstStubbing, ScalaOngoingStubbing }
 
-import scala.reflect.ClassTag
 import scala.reflect.macros.blackbox
 
 object WhenMacro {
+  // Re-export runtime classes so existing code referencing WhenMacro.* continues to work
+  type AnswerActions[T]   = WhenMacroRuntime.AnswerActions[T]
+  type AnswerPFActions[T] = WhenMacroRuntime.AnswerPFActions[T]
+  val RealMethod = WhenMacroRuntime.RealMethod
+
   private def transformInvocation(c: blackbox.Context)(invocation: c.Tree): c.Tree = {
     import c.universe.*
 
@@ -84,10 +87,6 @@ object WhenMacro {
     r
   }
 
-  object RealMethod {
-    def willBe(called: Called.type): Called.type = called
-  }
-
   val ShouldCallOptions                                                                                                          = Set("shouldCall", "mustCall", "calls")
   def shouldCallRealMethod[T: c.WeakTypeTag](c: blackbox.Context)(crm: c.Expr[RealMethod.type]): c.Expr[ScalaOngoingStubbing[T]] = {
     import c.universe.*
@@ -126,330 +125,6 @@ object WhenMacro {
     r
   }
 
-  class AnswerActions[T](os: ScalaFirstStubbing[T]) {
-    def apply(f: => T): ScalaOngoingStubbing[T] = os thenAnswer f
-
-    def apply[P0: ValueClassWrapper](f: P0 => T)(implicit classTag: ClassTag[P0] = defaultClassTag[P0]): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[P0: ValueClassWrapper, P1: ValueClassWrapper](f: (P0, P1) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[P0: ValueClassWrapper, P1: ValueClassWrapper, P2: ValueClassWrapper](f: (P0, P1, P2) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[P0: ValueClassWrapper, P1: ValueClassWrapper, P2: ValueClassWrapper, P3: ValueClassWrapper](f: (P0, P1, P2, P3) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[P0: ValueClassWrapper, P1: ValueClassWrapper, P2: ValueClassWrapper, P3: ValueClassWrapper, P4: ValueClassWrapper](
-        f: (P0, P1, P2, P3, P4) => T
-    ): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[P0: ValueClassWrapper, P1: ValueClassWrapper, P2: ValueClassWrapper, P3: ValueClassWrapper, P4: ValueClassWrapper, P5: ValueClassWrapper](
-        f: (P0, P1, P2, P3, P4, P5) => T
-    ): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[P0: ValueClassWrapper, P1: ValueClassWrapper, P2: ValueClassWrapper, P3: ValueClassWrapper, P4: ValueClassWrapper, P5: ValueClassWrapper, P6: ValueClassWrapper](
-        f: (P0, P1, P2, P3, P4, P5, P6) => T
-    ): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper,
-        P16: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15, P16) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper,
-        P16: ValueClassWrapper,
-        P17: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15, P16, P17) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper,
-        P16: ValueClassWrapper,
-        P17: ValueClassWrapper,
-        P18: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15, P16, P17, P18) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper,
-        P16: ValueClassWrapper,
-        P17: ValueClassWrapper,
-        P18: ValueClassWrapper,
-        P19: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15, P16, P17, P18, P19) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper,
-        P16: ValueClassWrapper,
-        P17: ValueClassWrapper,
-        P18: ValueClassWrapper,
-        P19: ValueClassWrapper,
-        P20: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15, P16, P17, P18, P19, P20) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-    def apply[
-        P0: ValueClassWrapper,
-        P1: ValueClassWrapper,
-        P2: ValueClassWrapper,
-        P3: ValueClassWrapper,
-        P4: ValueClassWrapper,
-        P5: ValueClassWrapper,
-        P6: ValueClassWrapper,
-        P7: ValueClassWrapper,
-        P8: ValueClassWrapper,
-        P9: ValueClassWrapper,
-        P10: ValueClassWrapper,
-        P11: ValueClassWrapper,
-        P12: ValueClassWrapper,
-        P13: ValueClassWrapper,
-        P14: ValueClassWrapper,
-        P15: ValueClassWrapper,
-        P16: ValueClassWrapper,
-        P17: ValueClassWrapper,
-        P18: ValueClassWrapper,
-        P19: ValueClassWrapper,
-        P20: ValueClassWrapper,
-        P21: ValueClassWrapper
-    ](f: (P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12, P13, P14, P15, P16, P17, P18, P19, P20, P21) => T): ScalaOngoingStubbing[T] =
-      os thenAnswer f
-
-//      (2 to 22).foreach { fn =>
-//        val args = 0 until fn
-//        print(s"""
-//                 |def apply[${args.map(a => s"P$a: ValueClassWrapper").mkString(",")}](f: (${args.map(a => s"P$a").mkString(",")}) => T): ScalaOngoingStubbing[T] =
-//                 |    os thenAnswer f
-//                 |""".stripMargin)
-//      }
-  }
-
   private val ShouldAnswerOptions                                 = Set("shouldAnswer", "mustAnswer", "answers")
   private val FunctionalShouldAnswerOptions                       = ShouldAnswerOptions.map(_ + "F")
   private val FunctionalShouldAnswerOptions2                      = ShouldAnswerOptions.map(_ + "FG")
@@ -458,7 +133,7 @@ object WhenMacro {
 
     val r = c.macroApplication match {
       case q"$_.StubbingOps[$t]($invocation).$m" if ShouldAnswerOptions.contains(m.toString) =>
-        q"new _root_.org.mockito.WhenMacro.AnswerActions(_root_.org.mockito.Mockito.when[$t](${transformInvocation(c)(invocation)}))"
+        q"new _root_.org.mockito.WhenMacroRuntime.AnswerActions(_root_.org.mockito.Mockito.when[$t](${transformInvocation(c)(invocation)}))"
 
       case q"$_.$cls[..$_]($invocation).$m" if cls.toString.startsWith("StubbingOps") && FunctionalShouldAnswerOptions.contains(m.toString) =>
         q"new _root_.org.mockito.${packageName(c)(cls)}.${className(c)(cls, "IdiomaticMockito")}.AnswerActions(_root_.org.mockito.Mockito.when(${transformInvocation(c)(invocation)}))"
@@ -472,17 +147,13 @@ object WhenMacro {
     r
   }
 
-  class AnswerPFActions[T](os: ScalaFirstStubbing[T]) {
-    def apply(pf: PartialFunction[Any, T]): ScalaOngoingStubbing[T] = os thenAnswer pf
-  }
-
   private val ShouldAnswerPFOptions                                 = Set("shouldAnswerPF", "mustAnswerPF", "answersPF")
   def shouldAnswerPF[T: c.WeakTypeTag](c: blackbox.Context): c.Tree = {
     import c.universe.*
 
     val r = c.macroApplication match {
       case q"$_.StubbingOps[$t]($invocation).$m" if ShouldAnswerPFOptions.contains(m.toString) =>
-        q"new _root_.org.mockito.WhenMacro.AnswerPFActions(_root_.org.mockito.Mockito.when[$t](${transformInvocation(c)(invocation)}))"
+        q"new _root_.org.mockito.WhenMacroRuntime.AnswerPFActions(_root_.org.mockito.Mockito.when[$t](${transformInvocation(c)(invocation)}))"
 
       case o => throw new Exception(s"Couldn't recognize ${show(o)}")
     }
